requirements.txt
fastapi
pydantic
uvicorn
motor
etcd3
protobuf==3.18.3
python-dotenv